adding a pytest for proposed sweep function
These tests ensure that the output results of the original and revised
functions are of the same length and all elements are identical to
within specified tolerances. It also checks to ensure that there is not
a significant slowdown compared to the original. This latter test is
stochastic due to how small these benchmarks are, and as a result, we
cannot guarantee strong parity at low Ns