moving particle exactly on boundary inside galaxy
At present, we check to see if the particle is outside the disk radius,
and if so, kick it back inside. We then check to ensure that all
particles are within the radius disk. However, this leads to a rare bug
where the particle is exactly on the disk radius, so it is not kicked
back in, but also fails the assert check. This commit changes the kick
logic to trigger when the particle is outside or exactly on the disk
radius